add silu test

+++ torch_dipu/csrc_dipu/aten/ops/CustomFallbackFunctions.hpp
@@ -17,7 +17,7 @@ at::Tensor to_cpu_no_half(const at::Tensor& devtensor) {
       return cpu_tensor.to(at::ScalarType::Float);
     } else {
       return cpu_tensor;
-    }  
+    }
 }
 
 at::Tensor& custom_fallback_dipu_silu_out(const at::Tensor& self, at::Tensor& out) {
@@ -26,7 +26,7 @@ at::Tensor& custom_fallback_dipu_silu_out(const at::Tensor& self, at::Tensor& ou
   auto self_cpu = to_cpu_no_half(self);
   auto out_cpu = to_cpu_no_half(self);
   out_cpu = at::silu_out(self_cpu, out_cpu);
-  out.copy_(out_cpu.to(at::ScalarType::Half));
+  out.copy_(out_cpu);
   return out;
 }
